Move GroupsUtil and MoveGroupChange to gui
