+++ src/main/java/net/sf/jabref/gui/groups/AutoGroupDialog.java
@@ -5,7 +5,11 @@ import java.awt.event.ActionEvent;
 import java.awt.event.ActionListener;
 import java.util.ArrayList;
 import java.util.List;
+import java.util.Optional;
 import java.util.Set;
+import java.util.StringTokenizer;
+import java.util.TreeSet;
+import java.util.stream.Collectors;
 
 import javax.swing.AbstractAction;
 import javax.swing.ActionMap;
@@ -29,16 +33,20 @@ import net.sf.jabref.gui.keyboard.KeyBinding;
 import net.sf.jabref.gui.undo.NamedCompound;
 import net.sf.jabref.logic.l10n.Localization;
 import net.sf.jabref.logic.layout.format.LatexToUnicodeFormatter;
+import net.sf.jabref.model.database.BibDatabase;
+import net.sf.jabref.model.entry.Author;
+import net.sf.jabref.model.entry.AuthorList;
+import net.sf.jabref.model.entry.BibEntry;
 import net.sf.jabref.model.entry.FieldName;
 import net.sf.jabref.model.groups.ExplicitGroup;
 import net.sf.jabref.model.groups.GroupHierarchyType;
 import net.sf.jabref.model.groups.GroupTreeNode;
-import net.sf.jabref.model.groups.GroupsUtil;
 
 import com.jgoodies.forms.builder.ButtonBarBuilder;
 import com.jgoodies.forms.builder.FormBuilder;
 import com.jgoodies.forms.layout.FormLayout;
 import net.sf.jabref.model.groups.SimpleKeywordGroup;
+import net.sf.jabref.model.strings.StringUtil;
 
 /**
  * Dialog for creating or modifying groups. Operates directly on the Vector containing group information.
@@ -85,21 +93,21 @@ class AutoGroupDialog extends JDialog implements CaretListener {
                 String fieldText = field.getText().toLowerCase().trim();
                 if (this.keywords.isSelected()) {
                     if (useCustomDelimiter.isSelected()) {
-                        keywords = GroupsUtil.findDeliminatedWordsInField(panel.getDatabase(), fieldText,
+                        keywords = findDeliminatedWordsInField(panel.getDatabase(), fieldText,
                                 deliminator.getText());
                     } else {
-                        keywords = GroupsUtil.findAllWordsInField(panel.getDatabase(), fieldText, remove.getText());
+                        keywords = findAllWordsInField(panel.getDatabase(), fieldText, remove.getText());
 
                     }
                 } else if (authors.isSelected()) {
                     List<String> fields = new ArrayList<>(2);
                     fields.add(FieldName.AUTHOR);
-                    keywords = GroupsUtil.findAuthorLastNames(panel.getDatabase(), fields);
+                    keywords = findAuthorLastNames(panel.getDatabase(), fields);
                     fieldText = FieldName.AUTHOR;
                 } else { // editors.isSelected() as it is a radio button group.
                     List<String> fields = new ArrayList<>(2);
                     fields.add(FieldName.EDITOR);
-                    keywords = GroupsUtil.findAuthorLastNames(panel.getDatabase(), fields);
+                    keywords = findAuthorLastNames(panel.getDatabase(), fields);
                     fieldText = FieldName.EDITOR;
                 }
 
@@ -181,6 +189,67 @@ class AutoGroupDialog extends JDialog implements CaretListener {
         updateComponents();
         pack();
         setLocationRelativeTo(frame);
+    }
+
+    public static Set<String> findDeliminatedWordsInField(BibDatabase db, String field, String deliminator) {
+        Set<String> res = new TreeSet<>();
+
+        for (BibEntry be : db.getEntries()) {
+            be.getField(field).ifPresent(fieldValue -> {
+                StringTokenizer tok = new StringTokenizer(fieldValue.trim(), deliminator);
+                while (tok.hasMoreTokens()) {
+                    res.add(StringUtil.capitalizeFirst(tok.nextToken().trim()));
+                }
+            });
+        }
+        return res;
+    }
+
+    /**
+     * Returns a Set containing all words used in the database in the given field type. Characters in
+     * <code>remove</code> are not included.
+     *
+     * @param db a <code>BibDatabase</code> value
+     * @param field a <code>String</code> value
+     * @param remove a <code>String</code> value
+     * @return a <code>Set</code> value
+     */
+    public static Set<String> findAllWordsInField(BibDatabase db, String field, String remove) {
+        Set<String> res = new TreeSet<>();
+        for (BibEntry be : db.getEntries()) {
+            be.getField(field).ifPresent(o -> {
+                StringTokenizer tok = new StringTokenizer(o, remove, false);
+                while (tok.hasMoreTokens()) {
+                    res.add(StringUtil.capitalizeFirst(tok.nextToken().trim()));
+                }
+            });
+        }
+        return res;
+    }
+
+    /**
+     * Finds all authors' last names in all the given fields for the given database.
+     *
+     * @param db The database.
+     * @param fields The fields to look in.
+     * @return a set containing the names.
+     */
+    public static Set<String> findAuthorLastNames(BibDatabase db, List<String> fields) {
+        Set<String> res = new TreeSet<>();
+        for (BibEntry be : db.getEntries()) {
+            for (String field : fields) {
+                be.getField(field).ifPresent(val -> {
+                    if (!val.isEmpty()) {
+                        AuthorList al = AuthorList.parse(val);
+                        res.addAll(al.getAuthors().stream().map(Author::getLast).filter(Optional::isPresent)
+                                .map(Optional::get).filter(lastName -> !lastName.isEmpty())
+                                .collect(Collectors.toList()));
+                    }
+                });
+            }
+        }
+
+        return res;
     }
 
     @Override
