+++ src/main/java/net/sf/jabref/gui/groups/GroupSelector.java
@@ -63,7 +63,6 @@ import net.sf.jabref.model.groups.AbstractGroup;
 import net.sf.jabref.model.groups.AllEntriesGroup;
 import net.sf.jabref.model.groups.EntriesGroupChange;
 import net.sf.jabref.model.groups.GroupTreeNode;
-import net.sf.jabref.model.groups.MoveGroupChange;
 import net.sf.jabref.model.groups.event.GroupUpdatedEvent;
 import net.sf.jabref.model.metadata.MetaData;
 import net.sf.jabref.model.search.SearchMatcher;
