+++ src/main/java/net/sf/jabref/gui/groups/UndoableMoveGroup.java
@@ -6,7 +6,6 @@ import java.util.Objects;
 import net.sf.jabref.gui.undo.AbstractUndoableJabRefEdit;
 import net.sf.jabref.logic.l10n.Localization;
 import net.sf.jabref.model.groups.GroupTreeNode;
-import net.sf.jabref.model.groups.MoveGroupChange;
 
 /**
  * @author jzieren
